Question2/requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy
pillow
matplotlib
scikit-learn
tqdm
streamlit
